+++ src/main/java/AMS/AMSsideproject/web/responseDto/post/PostListDtoAboutSpecificUser.java
@@ -25,7 +25,7 @@ public class PostListDtoAboutSpecificUser {
     @ApiModelProperty(example = "10")
     private Integer replyNum; //좋아요수
     @ApiModelProperty(example = "6")
-    private Long countView; //조회수
+    private Integer countView; //조회수
 
 
     @ApiModelProperty(example = "DFS", notes = "문제 태그들")
